requirements.txt
Django==4.2.11
django-widget-tweaks==1.5.0
djangorestframework==3.15.1

